Move UI templates into ui directory
